+++ memories.h
@@ -26,7 +26,7 @@ struct Memory {
     // chip VCC enable pin
     uint8_t PIN_VCC;
     // number of pages in device
-    const uint32_t page_num;
+    const uint16_t page_num;
     // size of one page in bytes
     const uint16_t page_size;
     // The port that the memory chip select is connected to
